+++ public/admin/_app/immutable/nodes/27.BDP1Rmzk.js
@@ -1,0 +1,1 @@
+import"../chunks/Dqbbk3fc.js";import"../chunks/xihTtKlq.js";import"../chunks/BkJzZCwM.js";import{t as e}from"../chunks/B2K3XSRK.js";function t(t){e(t,{})}export{t as component};